requirements.txt
# Core
numpy>=1.24
pandas>=2.0
scipy>=1.10
pyyaml>=6.0.1
matplotlib>=3.8
tqdm

# ML
scikit-learn>=1.3
xgboost>=2.0
lightgbm>=4.0

# Deep Learning
torch>=2.2
torchvision>=0.17

# Optimization & Logging
optuna>=3.5
tensorboard>=2.14

# Serving
fastapi>=0.111
uvicorn[standard]>=0.30
kafka-python>=2.0
python-dotenv

#data
yfinance
requests

akshare>=1.11.98
APScheduler>=3.10.4
jinja2>=3.1.3
plotly>=5.24.1
uvicorn[standard]>=0.30.0

#mysql
SQLAlchemy>=2.0.0
pymysql>=1.1.0
